alternative implementation of the sync.Map so the register is also buildable on go1.8
