+++ register/internal/untyped_other.go
@@ -1,0 +1,23 @@
+// +build !go1.9
+
+package internal
+
+import "sync"
+
+type Untyped struct {
+	data  map[string]interface{}
+	mutex *sync.RWMutex
+}
+
+func (u *Untyped) Register(name string, v interface{}) {
+	u.mutex.Lock()
+	u.data[name] = v
+	u.mutex.Unlock()
+}
+
+func (u *Untyped) Get(name string) (interface{}, bool) {
+	u.mutex.RLock()
+	v, ok := u.data[name]
+	u.mutex.RUnlock()
+	return v, ok
+}
